requirements.txt
youtube-transcript-api
python-dotenv
streamlit